refactor: improve name convention
